Refactor TerraformValidator so each ActionHandler has its own validator
