--- source/Sashimi.Terraform/Validation/TerraformDeploymentActionValidation.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using FluentValidation;
-using Sashimi.Server.Contracts.ActionHandlers.Validation;
-using Sashimi.Server.Contracts.CloudTemplates;
-
-namespace Sashimi.Terraform.Validation
-{
-    class TerraformDeploymentActionValidation : IDeploymentActionValidator
-    {
-        readonly ICloudTemplateHandlerFactory cloudTemplateHandlerFactory;
-
-        public TerraformDeploymentActionValidation(ICloudTemplateHandlerFactory cloudTemplateHandlerFactory)
-        {
-            this.cloudTemplateHandlerFactory = cloudTemplateHandlerFactory;
-        }
-
-        public void AddDeploymentValidationRule(AbstractValidator<DeploymentActionValidationContext> validator)
-        {
-            validator.Include(new TerraformValidator(cloudTemplateHandlerFactory));
-        }
-    }
-}
