+++ source/Sashimi.Terraform/Validation/TerraformPlanActionValidator.cs
@@ -1,0 +1,12 @@
+﻿using Sashimi.Server.Contracts.CloudTemplates;
+
+namespace Sashimi.Terraform.Validation
+{
+    class TerraformPlanActionValidator : TerraformDeploymentActionValidator
+    {
+        public TerraformPlanActionValidator(ICloudTemplateHandlerFactory cloudTemplateHandlerFactory)
+            : base(cloudTemplateHandlerFactory)
+        {
+        }
+    }
+}
