+++ source/Sashimi.Terraform/Validation/TerraformPlanDestroyActionValidator.cs
@@ -1,0 +1,12 @@
+﻿using Sashimi.Server.Contracts.CloudTemplates;
+
+namespace Sashimi.Terraform.Validation
+{
+    class TerraformPlanDestroyActionValidator : TerraformDeploymentActionValidator
+    {
+        public TerraformPlanDestroyActionValidator(ICloudTemplateHandlerFactory cloudTemplateHandlerFactory)
+            : base(cloudTemplateHandlerFactory)
+        {
+        }
+    }
+}
